TE-3870:
* Refactored tests
* Fixed typehits
