+++ tests/SprykerTest/Client/RabbitMq/Model/Connection/ConnectionManagerTest.php
@@ -8,17 +8,6 @@
 namespace SprykerTest\Client\RabbitMq\Model\Connection;
 
 use Codeception\Test\Unit;
-use Generated\Shared\Transfer\QueueConnectionTransfer;
-use Generated\Shared\Transfer\StoreTransfer;
-use PhpAmqpLib\Channel\AMQPChannel;
-use Spryker\Client\RabbitMq\Dependency\Client\RabbitMqToStoreClientBridge;
-use Spryker\Client\RabbitMq\Model\Connection\Connection;
-use Spryker\Client\RabbitMq\Model\Connection\ConnectionFactory;
-use Spryker\Client\RabbitMq\Model\Connection\ConnectionManager\ConnectionConfigFilter\ConnectionConfigFilter;
-use Spryker\Client\RabbitMq\Model\Connection\ConnectionManager\ConnectionConfigMapper\ConnectionConfigMapper;
-use Spryker\Client\RabbitMq\Model\Connection\ConnectionManager\ConnectionCreator\ConnectionCreator;
-use Spryker\Client\RabbitMq\Model\Connection\ConnectionManager\ConnectionManager;
-use Spryker\Client\RabbitMq\Model\Connection\ConnectionManager\ConnectionManagerInterface;
 use Spryker\Client\RabbitMq\Model\Exception\ConnectionConfigIsNotDefinedException;
 use Spryker\Client\RabbitMq\Model\Exception\DefaultConnectionNotFoundException;
 
@@ -37,7 +26,6 @@ class ConnectionManagerTest extends Unit
     protected const STORE_NAME = 'DE';
     protected const LOCALE_CODE = 'en_US';
     protected const QUEUE_POOL_NAME = 'synchronizationPool';
-    protected const DEFAULT_POOL_CONNECTION_NAME = self::STORE_NAME . '-name';
 
     protected const INCORRECT_LOCALE_CODE = 'INCORRECT_LOCALE_CODE';
     protected const INCORRECT_CONNECTION_NAME = 'INCORRECT_CONNECTION_NAME';
@@ -54,146 +42,9 @@ class ConnectionManagerTest extends Unit
     {
         $this->expectException(ConnectionConfigIsNotDefinedException::class);
 
-        $connectionManager = $this->createConnectionManager($this->createDefaultQueueConnectionTransfer(), static::INCORRECT_CONNECTION_NAME);
+        $connectionManager = $this->tester->createConnectionManagerWithDefaultQueueConnectionAndIncorrectLocale();
 
         $connectionManager->getChannelsByQueuePoolName(static::QUEUE_POOL_NAME, static::LOCALE_CODE);
-    }
-
-    /**
-     * @param string $storeName
-     *
-     * @return \Generated\Shared\Transfer\QueueConnectionTransfer
-     */
-    protected function createDefaultQueueConnectionTransfer(string $storeName = self::STORE_NAME): QueueConnectionTransfer
-    {
-        return $this->createQueueConnectionTransfer($storeName)
-            ->setIsDefaultConnection(true);
-    }
-
-    /**
-     * @param string $storeName
-     *
-     * @return \Generated\Shared\Transfer\QueueConnectionTransfer
-     */
-    protected function createQueueConnectionTransfer(string $storeName = self::STORE_NAME): QueueConnectionTransfer
-    {
-        return (new QueueConnectionTransfer())
-            ->setName(sprintf('%s-name', $storeName))
-            ->setHost(sprintf('%s-host', $storeName))
-            ->setPort(10)
-            ->setUsername(sprintf('%s-username', $storeName))
-            ->setPassword(sprintf('%s-password', $storeName))
-            ->setVirtualHost(sprintf('\%s-virtual-host', $storeName))
-            ->setStoreNames(['DE', 'AT']);
-    }
-
-    /**
-     * @param \Generated\Shared\Transfer\QueueConnectionTransfer $connectionTransfer
-     * @param string $poolConnectionName
-     *
-     * @return \Spryker\Client\RabbitMq\Model\Connection\ConnectionManager\ConnectionManagerInterface
-     */
-    protected function createConnectionManager(
-        QueueConnectionTransfer $connectionTransfer,
-        string $poolConnectionName = self::DEFAULT_POOL_CONNECTION_NAME
-    ): ConnectionManagerInterface {
-        $storeClient = $this->getStoreClientMock($poolConnectionName);
-        $connectionFactory = $this->getConnectionFactoryMock([$connectionTransfer]);
-
-        return new ConnectionManager(
-            $storeClient,
-            $connectionFactory,
-            new ConnectionConfigMapper($connectionFactory),
-            new ConnectionConfigFilter($storeClient, $connectionFactory),
-            new ConnectionCreator($connectionFactory)
-        );
-    }
-
-    /**
-     * @param string $poolConnectionName
-     *
-     * @return \Spryker\Client\RabbitMq\Dependency\Client\RabbitMqToStoreClientInterface|\PHPUnit\Framework\MockObject\MockObject
-     */
-    protected function getStoreClientMock(string $poolConnectionName)
-    {
-        $storeClientMock = $this->getMockBuilder(RabbitMqToStoreClientBridge::class)
-            ->disableOriginalConstructor()
-            ->getMock();
-
-        $storeClientMock
-            ->method('getCurrentStore')
-            ->willReturn($this->createStoreTransfer($poolConnectionName));
-
-        $storeClientMock
-            ->method('getStoreByName')
-            ->willReturn($this->createStoreTransfer($poolConnectionName));
-
-        return $storeClientMock;
-    }
-
-    /**
-     * @param string $poolConnectionName
-     *
-     * @return \Generated\Shared\Transfer\StoreTransfer
-     */
-    protected function createStoreTransfer(string $poolConnectionName): StoreTransfer
-    {
-        return (new StoreTransfer())
-            ->setQueuePools([
-                static::QUEUE_POOL_NAME => [
-                    $poolConnectionName,
-                ],
-            ])
-            ->setAvailableLocaleIsoCodes([
-                static::LOCALE_CODE,
-            ]);
-    }
-
-    /**
-     * @param \Generated\Shared\Transfer\QueueConnectionTransfer[] $queueConnectionTransfers
-     *
-     * @return \Spryker\Client\RabbitMq\Model\Connection\ConnectionFactoryInterface|\PHPUnit\Framework\MockObject\MockObject
-     */
-    protected function getConnectionFactoryMock(array $queueConnectionTransfers)
-    {
-        $connectionFactoryMock = $this->getMockBuilder(ConnectionFactory::class)
-            ->getMock();
-
-        $connectionFactoryMock
-            ->method('getQueueConnectionConfigs')
-            ->willReturn($queueConnectionTransfers);
-
-        $connectionFactoryMock
-            ->method('createConnection')
-            ->willReturn($this->getConnectionMock());
-
-        return $connectionFactoryMock;
-    }
-
-    /**
-     * @return \Spryker\Client\RabbitMq\Model\Connection\ConnectionInterface|\PHPUnit\Framework\MockObject\MockObject
-     */
-    protected function getConnectionMock()
-    {
-        $connectionMock = $this->getMockBuilder(Connection::class)
-            ->disableOriginalConstructor()
-            ->getMock();
-
-        $connectionMock
-            ->method('getChannel')
-            ->willReturn($this->createAMQPChannelMock());
-
-        return $connectionMock;
-    }
-
-    /**
-     * @return \PhpAmqpLib\Channel\AMQPChannel|\PHPUnit\Framework\MockObject\MockObject
-     */
-    protected function createAMQPChannelMock()
-    {
-        return $this->getMockBuilder(AMQPChannel::class)
-            ->disableOriginalConstructor()
-            ->getMock();
     }
 
     /**
@@ -203,7 +54,7 @@ class ConnectionManagerTest extends Unit
     {
         $this->expectException(DefaultConnectionNotFoundException::class);
 
-        $connectionManager = $this->createConnectionManager($this->createQueueConnectionTransfer());
+        $connectionManager = $this->tester->createConnectionManagerWithoutDefaultQueueConnection();
 
         $connectionManager->getDefaultChannel();
     }
@@ -213,7 +64,7 @@ class ConnectionManagerTest extends Unit
      */
     public function testGetChannelsByStoreNameForIncorrectLocale(): void
     {
-        $connectionManager = $this->createConnectionManager($this->createDefaultQueueConnectionTransfer());
+        $connectionManager = $this->tester->createConnectionManagerWithDefaultQueueConnection();
 
         $channels = $connectionManager
             ->getChannelsByStoreName(static::STORE_NAME, static::INCORRECT_LOCALE_CODE);
@@ -226,7 +77,7 @@ class ConnectionManagerTest extends Unit
      */
     public function testGetChannelsByQueuePoolNameWithoutLocale(): void
     {
-        $connectionManager = $this->createConnectionManager($this->createDefaultQueueConnectionTransfer());
+        $connectionManager = $this->tester->createConnectionManagerWithDefaultQueueConnection();
 
         $channels = $connectionManager->getChannelsByQueuePoolName(static::QUEUE_POOL_NAME, null);
 
@@ -238,7 +89,7 @@ class ConnectionManagerTest extends Unit
      */
     public function testGetChannelsByStoreNameForDefinedLocale(): void
     {
-        $connectionManager = $this->createConnectionManager($this->createDefaultQueueConnectionTransfer());
+        $connectionManager = $this->tester->createConnectionManagerWithDefaultQueueConnection();
 
         $channels = $connectionManager->getChannelsByStoreName(static::STORE_NAME, static::LOCALE_CODE);
 
@@ -250,7 +101,7 @@ class ConnectionManagerTest extends Unit
      */
     public function testGetChannelsByStoreNameWithoutLocale(): void
     {
-        $connectionManager = $this->createConnectionManager($this->createDefaultQueueConnectionTransfer());
+        $connectionManager = $this->tester->createConnectionManagerWithDefaultQueueConnection();
 
         $channels = $connectionManager->getChannelsByStoreName(static::STORE_NAME, null);
 
@@ -262,7 +113,7 @@ class ConnectionManagerTest extends Unit
      */
     public function testGetDefaultChannelShouldGetTheSameConnectionForMultipleCalls(): void
     {
-        $connectionManager = $this->createConnectionManager($this->createDefaultQueueConnectionTransfer());
+        $connectionManager = $this->tester->createConnectionManagerWithDefaultQueueConnection();
 
         $firstChannelResult = $connectionManager->getDefaultChannel();
         $secondChannelResult = $connectionManager->getDefaultChannel();
@@ -275,7 +126,7 @@ class ConnectionManagerTest extends Unit
      */
     public function testGetChannelsByStoreNameShouldGetTheSameConnectionForMultipleCalls(): void
     {
-        $connectionManager = $this->createConnectionManager($this->createDefaultQueueConnectionTransfer());
+        $connectionManager = $this->tester->createConnectionManagerWithDefaultQueueConnection();
 
         $firstChannelsResult = $connectionManager->getChannelsByStoreName(static::STORE_NAME, null);
         $secondChannelsResult = $connectionManager->getChannelsByStoreName(static::STORE_NAME, null);
@@ -288,7 +139,7 @@ class ConnectionManagerTest extends Unit
      */
     public function testGetChannelsByQueuePoolNameShouldGetTheSameConnectionForMultipleCalls(): void
     {
-        $connectionManager = $this->createConnectionManager($this->createDefaultQueueConnectionTransfer());
+        $connectionManager = $this->tester->createConnectionManagerWithDefaultQueueConnection();
 
         $firstChannelsResult = $connectionManager
             ->getChannelsByQueuePoolName(static::QUEUE_POOL_NAME, static::LOCALE_CODE);
