+++ tests/SprykerTest/Client/RabbitMq/codeception.yml
@@ -21,3 +21,4 @@ suites:
             - \SprykerTest\Shared\Testify\Helper\Environment
             - \SprykerTest\Shared\Testify\Helper\ConfigHelper
             - \SprykerTest\Shared\Testify\Helper\LocatorHelper
+            - \SprykerTest\Client\RabbitMq\Helper\RabbitMqHelper
